profiling/mock/pyproject.toml
[project]
name = "moccasin_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mamushi>=0.0.3",
    "snekmate>=0.1.0",
    "titanoboa>=0.2.4",
    "vyper>=0.4.0",
]
